Refactor graph controls out of prefigure
Make graph controls renderer-agnostic and move control UI into GraphControls.

Add renderer-agnostic worker and Cypress coverage; restore GraphControls documentation.
